accel/amdxdna: add amdxdna_client_populate_ranges()
Add a client-level helper that safely drains invalid BOs by repopulating
their ranges. Keep the HMM faulting path out of AUX builds because VE2
does not submit through the AIE2/AIE4 paths and its kernel does not export
hmm_range_fault().

(cherry picked from commit e95c5a45fecd51d7d5c5ca51ec746390b50c4182)
